requirements.txt
absl-py==2.3.1
affine==2.4.0
aiohttp==3.13.2
aiosignal==1.4.0
albumentations==2.0.8
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
grpcio==1.75.1
h5py==3.15.1
huggingface-hub==0.19.4
hypll==0.1.1
idna==3.11
imagecodecs==2026.1.14
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
omegaconf
opencv-python-headless==4.12.0.88
pandas==2.3.3
pillow==11.3.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.12.2
six==1.17.0
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.20
torchmetrics==1.8.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
xformers==0.0.35
yacs==0.1.8
yarl==1.22.0